Added add/remove buttons to resource summaries
Icons could perhaps indicate that the item
is being added and remove from the map better.

The colours of the buttons will clash with the background.

The list of items should jump to the newly added item.
There should be some animation for removing items from the list.

There should be some sort of undo for adding/removing resources.